% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ons_sitemap.R
\name{convert_ons_taxonomy}
\alias{convert_ons_taxonomy}
\title{Convert ONS taxonomy data into a flat structure with additional columns}
\usage{
convert_ons_taxonomy(ons_taxonomy_all)
}
\arguments{
\item{ons_taxonomy_all}{A tibble containing the ONS taxonomy data}
}
\value{
A tibble with the ONS taxonomy data in a flat structure with additional columns
}
\description{
Convert ONS taxonomy data into a flat structure with additional columns
}
