% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ons_sitemap.R
\name{get_api_response_parsed}
\alias{get_api_response_parsed}
\title{Generate ONS taxonomy sitemap}
\usage{
get_api_response_parsed(req)
}
\arguments{
\item{base_url}{}
}
\description{
Generate ONS taxonomy sitemap
}
