% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ons_sitemap.R
\name{get_ons_taxonomy}
\alias{get_ons_taxonomy}
\title{Generate ONS taxonomy sitemap}
\usage{
get_ons_taxonomy(base_url = apis$ons$url$main)
}
\arguments{
\item{base_url}{The base URL of the ONS API.}
}
\value{
A sitemap for the ONS taxonomy API.
}
\description{
This function generates a sitemap for the ONS taxonomy API.
}
\examples{
get_ons_taxonomy("https://api.ons.gov.uk")
}
