% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ons_sitemap.R
\name{list_ons_datasets}
\alias{list_ons_datasets}
\title{Generate ONS taxonomy sitemap}
\usage{
list_ons_datasets(base_url = apis$ons$url$main)
}
\arguments{
\item{base_url}{}
}
\description{
Generate ONS taxonomy sitemap
}
