% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ons_sitemap.R
\name{list_ons_product_pages}
\alias{list_ons_product_pages}
\title{List ONS Product Pages}
\usage{
list_ons_product_pages(base_url = apis$ons$url$main)
}
\arguments{
\item{base_url}{}
}
\description{
List ONS Product Pages
}
