% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ons.R
\name{search_ons}
\alias{search_ons}
\title{Search the ONS API}
\usage{
search_ons(query, type = c("timeseries", "dataset"), limit = 100)
}
\arguments{
\item{query}{The search query string.}

\item{type}{The type(s) of items to search for. Must be one or more of "timeseries" or "dataset".}

\item{limit}{The maximum number of items to return.}
}
\value{
A list of search results.
}
\description{
This function searches the ONS API for timeseries or datasets based on a query string.
}
\examples{
search_ons("GDP")

}
\references{
\url{[ONS Zebedee Reader Search API]()}
}
\seealso{
\code{\link{get_ons_data}}, \code{\link{get_ons_metadata}}
}
